Pagamento Adapter e injeção de Pedidos e Pagamentos Use Case

+++ src/core/domain/repositories/cliente_repository.rs
@@ -8,7 +8,7 @@ pub trait ClienteRepository: Send + Sync{
 
     async fn get_cliente_by_cpf(&self, cpf: Cpf) -> Result<Cliente, DomainError>;
 
-    async fn get_cliente_by_id(&self, id: i32) -> Result<Cliente, DomainError>;
+    async fn get_cliente_by_id(&self, id: usize) -> Result<Cliente, DomainError>;
 
     async fn create_cliente(&mut self, cliente: Cliente) -> Result<Cliente, DomainError>;
